requirements.txt
appdirs==1.4.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
certifi==2017.4.17
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
Fuzzy==1.1
html5lib==0.999999999
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.0.1
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
nltk==3.2.2
notebook==5.0.0
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-Levenshtein==0.12.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
scandir==1.5
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==2.0.0
